refactor: move ` convertMarkdownToDiffChunk` to it's own file.
